+++ sdk/ai/Azure.AI.Projects/tests/AgentsTests.cs
@@ -5,10 +5,8 @@ using System.ClientModel;
 using System.ClientModel.Primitives;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
-using System.Diagnostics.Metrics;
 using System.IO;
 using System.Linq;
-using System.Net;
 using System.Text;
 using System.Text.Json;
 using System.Text.Json.Nodes;
@@ -1692,23 +1690,24 @@ public class AgentsTests : AgentsTestBase
             await Delay();
             session = await projectClient.AgentAdministrationClient.GetSessionAsync(agentName: agentVersion.Name, sessionId: session.AgentSessionId);
         }
-        HeaderTestPolicy sessionPolicy = new(new Dictionary<string, string>()
-        {
-            { "x-agent-session-id", session.AgentSessionId }
-        });
         ProjectOpenAIClientOptions responsesOptions = new()
         {
             Endpoint = new Uri(TestEnvironment.FOUNDRY_PROJECT_ENDPOINT),
             ApiVersion = "v1",
             AgentName = agentVersion.Name
         };
-        responsesOptions.AddPolicy(sessionPolicy, PipelinePosition.PerCall);
         responsesOptions = GetConfiguredOptions(
             responsesOptions,
             true);
         ProjectResponsesClient responseClient = CreateProxyFromClient(projectClient.ProjectOpenAIClient.GetProjectResponsesClientForAgentEndpoint(patchedRecord.Name, options: responsesOptions));
-        ResponseResult response = await responseClient.CreateResponseAsync("Hello, tell me a joke.");
-        Assert.That(response.GetOutputText(), Is.Not.Empty);
+        ProjectCreateResponseOptions responseOptions = new()
+        {
+            InputItems = { ResponseItem.CreateUserMessageItem("Hello, tell me a joke.") },
+            SessionId = session.AgentSessionId,
+        };
+        ClientResult<ResponseResult> response = await responseClient.CreateResponseAsync(responseOptions);
+        Assert.That(response.GetRawResponse().Headers, Does.Contain(new KeyValuePair<string, string>("x-agent-session-id", session.AgentSessionId)));
+        Assert.That(response.Value.GetOutputText(), Is.Not.Empty);
         // Disable the Agent
         await projectClient.AgentAdministrationClient.DisableAgentAsync(agentVersion.Name);
         try
@@ -1728,23 +1727,14 @@ public class AgentsTests : AgentsTestBase
             await Delay();
             session = await projectClient.AgentAdministrationClient.GetSessionAsync(agentName: agentVersion.Name, sessionId: session.AgentSessionId);
         }
-        sessionPolicy = new(new Dictionary<string, string>()
+        responseOptions = new()
         {
-            { "x-agent-session-id", session.AgentSessionId }
-        });
-        responsesOptions = new ProjectOpenAIClientOptions()
-        {
-            Endpoint = new Uri(TestEnvironment.FOUNDRY_PROJECT_ENDPOINT),
-            ApiVersion = "v1",
-            AgentName = agentVersion.Name
+            InputItems = { ResponseItem.CreateUserMessageItem("Hello, tell me a joke.") },
+            SessionId = session.AgentSessionId,
         };
-        responsesOptions.AddPolicy(sessionPolicy, PipelinePosition.PerCall);
-        responsesOptions = GetConfiguredOptions(
-            responsesOptions,
-            true);
-        responseClient = CreateProxyFromClient(projectClient.ProjectOpenAIClient.GetProjectResponsesClientForAgentEndpoint(patchedRecord.Name, options: responsesOptions));
-        response = await responseClient.CreateResponseAsync("Hello, tell me a joke.");
-        Assert.That(response.GetOutputText(), Is.Not.Empty);
+        response = await responseClient.CreateResponseAsync(responseOptions);
+        Assert.That(response.GetRawResponse().Headers, Does.Contain(new KeyValuePair<string, string>("x-agent-session-id", session.AgentSessionId)));
+        Assert.That(response.Value.GetOutputText(), Is.Not.Empty);
     }
 
     [RecordedTest]
@@ -1924,7 +1914,7 @@ public class AgentsTests : AgentsTestBase
               activity: hello world
             - kind: EndConversation
               id: end_conversation
-        """.ReplaceLineEndings("\r\n");
+        """.Replace("\r", "");
 
     private static async Task DeleteMemoryStoreMayBe(AIProjectClient projectClient, string name)
     {